Move PrioritySampling flags to dd-trace-api project.
