+++ dd-trace-api/src/main/java/datadog/trace/api/sampling/PrioritySampling.java
@@ -1,7 +1,5 @@
-package datadog.trace.common.sampling;
+package datadog.trace.api.sampling;
 
-/** @deprecated Replaced by {@link datadog.trace.api.sampling.PrioritySampling} . */
-@Deprecated
 public class PrioritySampling {
   /**
    * Implementation detail of the client. will not be sent to the agent or propagated.
