+++ dd-trace-ot/src/main/java/datadog/opentracing/DDTracer.java
@@ -13,10 +13,10 @@ import datadog.opentracing.scopemanager.ScopeContext;
 import datadog.trace.api.DDTags;
 import datadog.trace.api.interceptor.MutableSpan;
 import datadog.trace.api.interceptor.TraceInterceptor;
+import datadog.trace.api.sampling.PrioritySampling;
 import datadog.trace.common.DDTraceConfig;
 import datadog.trace.common.Service;
 import datadog.trace.common.sampling.AllSampler;
-import datadog.trace.common.sampling.PrioritySampling;
 import datadog.trace.common.sampling.RateByServiceSampler;
 import datadog.trace.common.sampling.Sampler;
 import datadog.trace.common.writer.DDAgentWriter;
