+++ dd-trace-ot/src/test/groovy/datadog/opentracing/SpanFactory.groovy
@@ -1,6 +1,6 @@
 package datadog.opentracing
 
-import datadog.trace.common.sampling.PrioritySampling
+import datadog.trace.api.sampling.PrioritySampling
 import datadog.trace.common.writer.ListWriter
 
 class SpanFactory {
