+++ dd-trace-ot/src/test/groovy/datadog/trace/api/sampling/RateByServiceSamplerTest.groovy
@@ -3,7 +3,6 @@ package datadog.trace.api.sampling
 import com.fasterxml.jackson.databind.ObjectMapper
 import datadog.opentracing.DDSpan
 import datadog.opentracing.SpanFactory
-import datadog.trace.common.sampling.PrioritySampling
 import datadog.trace.common.sampling.RateByServiceSampler
 import spock.lang.Specification
 import spock.lang.Timeout
